+++ lib/trackguard/adapters/local.rb
@@ -19,7 +19,9 @@ module Trackguard
         Visitor.flagged?(ip)
       end
 
-      def track_blocked_request(ip:, user_agent:, path:, http_method:, block_reason:)
+      protected
+
+      def perform_track_blocked_request(ip:, user_agent:, path:, http_method:, block_reason:)
         TrackBlockedRequestJob.perform_later(
           ip: ip,
           user_agent: user_agent,
@@ -28,8 +30,6 @@ module Trackguard
           block_reason: block_reason
         )
       end
-
-      protected
 
       def perform_track_page_view(path:, ip:, user_agent:, referer:, session_id:, trace_id:, source:,
                                   tracking_layer:, http_method:, prefetch: false)